backend-py/requirements.txt
fastapi
uvicorn
python-dotenv
structlog
